pyproject.toml
[project]
name = "robospec"
version = "0.1.0"
description = "Natural language to Isaac Lab RL environments"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
dependencies = [
    "typer[all]>=0.9.0",
    "httpx>=0.25.0",
    "pydantic>=2.5.0",
    "jinja2>=3.1.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
ui = ["streamlit>=1.29.0"]

[project.scripts]
robospec = "robospec.cli:app"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
